# api/api.yml
openapi: 3.0.3
info:
  title: Friends Net API
  description: |
    Frineds Net server API.
  version: 1.0.0

# Using Bearer HTTP authentication, can be switched to OAuth2, OpenId, ...
security:
   - JsonWebToken: []
# Tags are used to group endpoints. There can be multiple tags per endpoint,
# the first tag determines the interface (in case of Java) defining this endpoint
# in the generated code.
...
paths:
...
  '/api/users':
    get:
      summary: Searches for users
      description: Finds users who's name contains given substring.
      tags:
        - UserSearchController
      operationId: findUsers
      parameters:
       - $ref: '#/components/parameters/UserSearchParameter'
      responses:
       '200':
         $ref: '#/components/responses/UserSearchResponse'
       '401':
         description: Unauthenticated
  '/api/friendship/request':
    post:
      summary: Sends friend request
      description: Sends friend request to user with given username. They may accept, decline or block the request sender.
      tags:
        - UserRelationshipController
      operationId: createFriendRequest
      parameters:
        - $ref: '#/components/parameters/UsernameQueryParameter'
      responses:
        '200':
          description: OK. Friend request successfully sent.
        '400':
          description: User with given username does not exist.
        '401':
          description: Unauthenticated
    get:
      summary: Finds all friend requests
      description: Finds all pending friend requests for currently authenticated user.
      tags:
        - UserSearchController
      operationId: findFriendRequests
      responses:
        '200':
          $ref: '#/components/responses/UserSearchResponse'
        '401':
          description: Unauthenticated
...
  '/api/users/friends':
    get:
      summary: Finds user's friends.
      description: Finds all users that are friends with authenticated user.
      tags:
        - UserSearchController
      operationId: findFriends
      responses:
        '200':
          $ref: '#/components/responses/UserSearchResponse'
        '401':
          description: Unauthenticated
  '/api/users/blocked':
    get:
      summary: Finds blocked users
      description: Finds all users that authenticated user blocked.
      tags:
        - UserSearchController
      operationId: findBlockedUsers
      responses:
        '200':
          $ref: '#/components/responses/UserSearchResponse'
        '401':
          description: Unauthenticated
  '/api/users/friends-roles':
    get:
      summary: Finds user's friends with roles.
      description: Finds all users that are friends with authenticated user including their roles.
      tags:
        - UserSearchController
      operationId: findFriendsWithRoles
      responses:
        '200':
          $ref: '#/components/responses/UserWithRolesSearchResponse'
        '401':
          description: Unauthenticated
...
components:
  securitySchemes:
    JsonWebToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    UserSearchResponse:
      description: Response on successfull user search query.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FoundUsersVO'
    UserWithRolesSearchResponse:
      description: Response on successfull user with roles search query.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FoundUsersWithRolesVO'
...
  parameters:
    UserSearchParameter:
      name: nameLike
      in: query
      required: true
      description: Substring sought in user's names.
      allowEmptyValue: false
      schema:
        type: string
        minLength: 3
      example: Adam
    UsernameQueryParameter:
      name: username
      in: query
      required: true
      description: Username of the user we want to send friend request to.
      allowEmptyValue: false
      schema:
        type: string
        minLength: 4
      example: boss23@mail.com
...
  schemas:
...
    UserRelationshipVO:
      description: Object with data about found user and it's relationship status with searching user.
      allOf:
        - $ref: '#/components/schemas/UserIdentificationDataVO'
        - type: object
          properties:
            relationshipStatus:
              $ref: '#/components/schemas/EnumRelationshipStatus'
      required:
        - name
        - login
      example:
        name: John Doe
        login: john.doe@gmail.com
        relationshipStatus: BLOCKED
    EnumRelationshipStatus:
      description: Relationship status of two users.
      type: string
      enum:
        - REQUEST_SENT
        - FRIENDS
        - BLOCKED
...
    UserIdentificationDataVO:
      description: Object with user's identification data.
      properties:
        name:
          type: string
          description: User's name that will be displayed.
          maxLength: 50
          example: John Doe
        login:
          type: string
          format: email
          description: User's email. Will be used as login.
          maxLength: 50
          minLength: 4
          example: great@email.com
      example:
        name: John Doe
        login: great@email.com
      required:
        - name
        - login
    UserIdentificationDataWithRolesVO:
      description: Object with user's identification data and roles.
      allOf:
        - $ref: '#/components/schemas/UserIdentificationDataVO'
        - type: object
          properties:
            roles:
              type: array
              items:
                $ref: '#/components/schemas/EnumUserRole'
      required:
        - name
        - login
        - roles
...
    EnumUserRole:
      description: Enumeration of user roles
      type: string
      enum:
        - 'USER'
        - 'ADMIN'
    FoundUsersVO:
      description: Object carrying a list of found users.
      type: array
      items:
        $ref: '#/components/schemas/UserRelationshipVO'
    FoundUsersWithRolesVO:
      description: Object carrying a list of found users including their roles.
      type: array
      items:
        $ref: '#/components/schemas/UserIdentificationDataWithRolesVO'
...